1.修复BUF
2.优惠券接口开发中
3.租车接口逻辑变更，蓝牙锁在租车接口不生成订单，新增蓝牙订单接口
4.新增通过车牌号码获取车辆总共被租用次数和租用时间

+++ ASBicycle.Web/Controllers/Alipay/AlipayController.cs
@@ -309,8 +309,8 @@ namespace ASBicycle.Web.Controllers.Alipay
                             await _rechargeDetailWriteRepository.UpdateAsync(recharge_detail);
 
                             var recharge = await _rechargeWriteRepository.FirstOrDefaultAsync(t => t.User_id == recharge_detail.User_id);
-
-                            recharge.Recharge_count = (recharge.Recharge_count ?? 0) + recharge_detail.Recharge_amount; ;
+                            
+                            recharge.Recharge_count = (double)((decimal)(recharge.Recharge_count ?? 0) + (decimal)recharge_detail.Recharge_amount); ;
                             recharge.Updated_at = DateTime.Now;
 
                             await _rechargeWriteRepository.UpdateAsync(recharge);
